refactor: migrate optimize client usage

+++ optimize/util/optimize-commons/src/main/java/io/camunda/optimize/service/db/es/schema/ElasticSearchIndexSettingsBuilder.java
@@ -48,9 +48,10 @@ public class ElasticSearchIndexSettingsBuilder {
                         m.nestedObjects(
                             n ->
                                 n.limit(
-                                    configurationService
-                                        .getElasticSearchConfiguration()
-                                        .getNestedDocumentsLimit()))));
+                                    (long)
+                                        configurationService
+                                            .getElasticSearchConfiguration()
+                                            .getNestedDocumentsLimit()))));
   }
 
   public static IndexSettings buildAllSettings(
@@ -76,9 +77,10 @@ public class ElasticSearchIndexSettingsBuilder {
                       m.nestedObjects(
                           n ->
                               n.limit(
-                                  configurationService
-                                      .getElasticSearchConfiguration()
-                                      .getNestedDocumentsLimit())));
+                                  (long)
+                                      configurationService
+                                          .getElasticSearchConfiguration()
+                                          .getNestedDocumentsLimit())));
           try {
             indexMappingCreator.getStaticSettings(i, configurationService);
           } catch (final IOException e) {
